refactor(encryption): move encrypt-column into tables router
The POST /tables/:tableName/encrypt-column handler previously lived in
encryptionRouter mounted at /, which meant it squatted on the /tables
namespace owned by tablesRouter and only resolved via Express
fall-through. Move the handler into tables.routes.ts where it belongs
(table-scoped operation) and mount encryptionRouter at /encryption so
its remaining routes (GET / list, POST /rotate) are properly scoped.

External URLs are unchanged:
  POST /api/database/tables/:tableName/encrypt-column
  GET  /api/database/encryption
  POST /api/database/encryption/rotate
